Update scores when details view is loaded (if changed)
While using IITC mobile if you've just captured a portal you might want
to see the change to the score immediately, but previously landgrab
would only recalculate scores when it loads new portal data (typically
when you scroll or zoom). This patch makes it also recalculate when you
view portal details for a portal that is newly captured.

The logic for adding portals to updateQueue then popping them and
updating their scores doesn't work in this case so instead we just
rescore all portals that are currently loaded by intel (window.portals).
This is kinda nicer anyway; loading portals could change the score for
nearby portals so it's more accurate for that case too.

If capturing a portal changes the score for a portal that's currently
offscreen (because you have a huge capture radius) you'll still need to
scroll over that way to get it to update but this is still an
improvement.