Rewrite using Voronoi diagrams instead of circles
Major changes.

- all bubbles are gone.
- d3-delaunay is used to compute the Voronoi diagram for the portal
network.
- d3-delaunay only supports using integer indices for points, so we
can't use the ingress guid directly. To handle this, we introduce a
separate lookup table `landgrab.portalIndex` and take care to make sure
it stays in sync with `landgrab.portalInfo`. The latter is now just
a list and stores the guid as a property instead of being a dictionary
key.
- The new scoring system but it makes a lot of sense but it doesn't
reward players anywhere near as much for getting lots of portals in a
cluster. Possibly we can tweak it.