Highlight "blob" containing selected portal
Redrawing the voronoi diagrams outside of mapDataRefreshEnd means the
old strategy of setting landgrab.voronoi or landgrab.captureScores etc
to null to indicate they need to be regenerated is no good any more.

Instead we add "stale" flags for the voronoi diagram and the scores and
atomically swap in the new voronoi diagram / score object when they're
finished computing.

There might be a period where the voronoi diagram is out of sync with
the scores. Um. If that happens I'm sorry.