Fix for huge regions when portals are too close
When portals are very close together, computing the voronoi diagram goes
wrong, presumably because of rounding errors.

To work around this issue we store coords in the latE6, lngE6 units
ingress uses internally (i.e, the values multiplied by 1000000), compute
the voronoi diagram in those units and only convert back to actual
degrees when drawing the polygons.

To preserve backwards compatibility we add a new dataformat value in
localstorage so that we know to convert data to the new format.